requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
sqlite3>=3.35.0
